Refactored cleanup test cases
Refactored cleanup test cases to one file and deleted duplicated code.
